feat(help): restructure help command constants and add new command help interfaces refactor(kv): reorganize Redis client and key management functions fix(permissions): enhance channel admin check with creator validation
