+++ server/constants/help/ban.ts
@@ -1,0 +1,14 @@
+import type { CommandHelp } from '~/types';
+
+export const banHelp: CommandHelp = {
+  name: 'ban',
+  description: 'Ban a user from interacting with Gork.',
+  subcommands: [
+    {
+      usage: '/gork ban [@user ...]',
+      description:
+        'Ban one or more users. Opens a picker modal if no users specified.',
+      adminOnly: true,
+    },
+  ],
+};
