+++ server/constants/help/index.ts
@@ -1,0 +1,5 @@
+// biome-ignore lint/performance/noBarrelFile: intentional public API for help constants
+export { banHelp } from './ban';
+export { modeHelp } from './mode';
+export { reportsHelp } from './reports';
+export { unbanHelp } from './unban';
