+++ server/lib/kv/client.ts
@@ -1,0 +1,4 @@
+import { RedisClient } from 'bun';
+import { env } from '~/env';
+
+export const redis = new RedisClient(env.REDIS_URL);
