+++ server/lib/kv/index.ts
@@ -1,0 +1,11 @@
+// biome-ignore lint/performance/noBarrelFile: intentional public API for kv module
+export { redis } from './client';
+export { redisKeys } from './keys';
+export {
+  clearChannelMode,
+  getChannelMode,
+  type ResponseMode,
+  setChannelMode,
+} from './mode';
+export { ratelimit } from './ratelimit';
+export { clearSilenced, isSilenced, setSilenced } from './silence';
