+++ server/types/help.ts
@@ -1,0 +1,17 @@
+export interface HelpSubcommand {
+  adminOnly?: boolean;
+  description: string;
+  usage: string;
+}
+
+export interface HelpMode {
+  description: string;
+  name: string;
+}
+
+export interface CommandHelp {
+  description: string;
+  modes?: HelpMode[];
+  name: string;
+  subcommands: HelpSubcommand[];
+}
