+++ server/types/index.ts
@@ -1,4 +1,5 @@
 export * from '~/lib/validators/pinecone';
 export * from './activity';
+export * from './help';
 export * from './request';
 export * from './slack';
